requirements.txt
absl-py==0.14.0
asgiref==3.4.1
astunparse==1.6.3
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.6
clang==5.0
dj-database-url==0.5.0
Django==3.2.6
django-ckeditor==6.1.0
django-countries==7.2.1
django-filter==2.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.13.4
django-phonenumber-field==5.2.0
django-resized==0.3.11
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
gunicorn==20.1.0
h5py==3.1.0
idna==3.2
Markdown==3.3.4
mysqlclient==2.0.3
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
phonenumbers==8.12.39
phonenumberslite==8.12.35
Pillow==8.3.1
protobuf==3.18.0
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
whitenoise==5.3.0
wrapt==1.12.1
